requirements.txt
fastapi
sqlalchemy
passlib
pydantic
uvicorn
python-jose